Move dagger module out of autofill-store to sutofill-impl
