+++ autofill/autofill-store/build.gradle
@@ -1,14 +1,12 @@
 plugins {
     id 'com.android.library'
     id 'kotlin-android'
-    id 'com.squareup.anvil'
     id 'kotlin-android-extensions'
 }
 
 apply from: "$rootProject.projectDir/gradle/android-library.gradle"
 
 dependencies {
-    implementation project(path: ':di')
     implementation project(path: ':common')
     implementation project(path: ':autofill-api')
     implementation project(path: ':secure-storage-api')
@@ -24,21 +22,11 @@ dependencies {
 
     // Testing dependencies
     testImplementation Testing.junit4
+    testImplementation AndroidX.test.ext.junit
     testImplementation Testing.robolectric
     testImplementation (KotlinX.coroutines.test) {
         // https://github.com/Kotlin/kotlinx.coroutines/issues/2023
         // conflicts with mockito due to direct inclusion of byte buddy
         exclude group: "org.jetbrains.kotlinx", module: "kotlinx-coroutines-debug"
-    }
-}
-
-android {
-    anvil {
-        generateDaggerFactories = true // default is false
-    }
-    testOptions {
-        unitTests {
-            includeAndroidResources = true
-        }
     }
 }
